Move utils.tests package to src/test
