+++ src/test/java/org/magicdgs/readtools/CommandLineProgramTest.java
@@ -22,16 +22,9 @@
  * SOFTWARE.
  */
 
-package org.magicdgs.readtools.utils.tests;
+package org.magicdgs.readtools;
 
-import org.magicdgs.readtools.Main;
-
-import htsjdk.samtools.util.Log;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.broadinstitute.hellbender.utils.LoggingUtils;
 import org.broadinstitute.hellbender.utils.test.CommandLineProgramTester;
-import org.testng.annotations.BeforeSuite;
 
 import java.io.File;
 import java.util.ArrayList;
