+++ src/test/java/org/magicdgs/readtools/TestResourcesUtils.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.magicdgs.readtools.utils.tests;
+package org.magicdgs.readtools;
 
 import java.io.File;
