+++ src/test/java/org/magicdgs/readtools/cmd/argumentcollections/RTInputArgumentCollectionUnitTest.java
@@ -25,7 +25,7 @@
 package org.magicdgs.readtools.cmd.argumentcollections;
 
 import org.magicdgs.readtools.engine.RTDataSource;
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
