+++ src/test/java/org/magicdgs/readtools/cmd/argumentcollections/RTOutputBamArgumentCollectionUnitTest.java
@@ -25,7 +25,7 @@
 package org.magicdgs.readtools.cmd.argumentcollections;
 
 import org.magicdgs.readtools.exceptions.RTUserExceptions;
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import htsjdk.samtools.SAMFileHeader;
 import htsjdk.samtools.SAMProgramRecord;
