+++ src/test/java/org/magicdgs/readtools/cmd/argumentcollections/RTOutputFastqArgumentCollectionUnitTest.java
@@ -27,7 +27,7 @@ package org.magicdgs.readtools.cmd.argumentcollections;
 import org.magicdgs.readtools.utils.read.writer.FastqGATKWriter;
 import org.magicdgs.readtools.utils.read.writer.ReadToolsIOFormat;
 import org.magicdgs.readtools.utils.read.writer.SplitGATKWriter;
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import htsjdk.samtools.SAMFileHeader;
 import org.broadinstitute.hellbender.utils.read.GATKReadWriter;
