+++ src/test/java/org/magicdgs/readtools/tools/trimming/TrimReadsIntegrationTest.java
@@ -24,7 +24,7 @@
 
 package org.magicdgs.readtools.tools.trimming;
 
-import org.magicdgs.readtools.utils.tests.CommandLineProgramTest;
+import org.magicdgs.readtools.CommandLineProgramTest;
 
 import org.broadinstitute.hellbender.utils.test.ArgumentsBuilder;
 import org.broadinstitute.hellbender.utils.test.IntegrationTestSpec;
