+++ src/test/java/org/magicdgs/readtools/utils/distmap/DistmapGATKWriterUnitTest.java
@@ -25,7 +25,7 @@
 package org.magicdgs.readtools.utils.distmap;
 
 import org.magicdgs.readtools.utils.read.ReadWriterFactory;
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import org.broadinstitute.hellbender.exceptions.UserException;
 import org.broadinstitute.hellbender.utils.read.ArtificialReadUtils;
