+++ src/test/java/org/magicdgs/readtools/utils/iterators/ReadTransformerIteratorUnitTest.java
@@ -24,7 +24,7 @@
 
 package org.magicdgs.readtools.utils.iterators;
 
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import htsjdk.samtools.CigarOperator;
 import org.broadinstitute.hellbender.transformers.ReadTransformer;
