+++ src/test/java/org/magicdgs/readtools/utils/iterators/paired/GATKReadPairedIteratorUnitTest.java
@@ -24,7 +24,7 @@
 
 package org.magicdgs.readtools.utils.iterators.paired;
 
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import htsjdk.samtools.SAMFileHeader;
 import org.broadinstitute.hellbender.utils.read.ArtificialReadUtils;
