+++ src/test/java/org/magicdgs/readtools/utils/read/transformer/barcodes/FixRawBarcodeTagsReadTransformerUnitTest.java
@@ -24,7 +24,7 @@
 
 package org.magicdgs.readtools.utils.read.transformer.barcodes;
 
-import org.magicdgs.readtools.utils.tests.BaseTest;
+import org.magicdgs.readtools.BaseTest;
 
 import org.broadinstitute.hellbender.transformers.ReadTransformer;
 import org.broadinstitute.hellbender.utils.read.ArtificialReadUtils;
